- Refactor `UploadContext` to only pass `addFileToUploadQueue()` function   in its value - Rename `Upload` component to `DropZone` - Rename `Context` folder to `Reducer` and move `UploadContext` to parent   folder - Create `UploadManager` component and return as child of   `UploadContext.Provider`
