--- src/components/Upload/Context/index.ts
@@ -1,1 +1,0 @@
-export * from './UploadContext';
