+++ src/components/Upload/Dropzone.stories.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { Dropzone as DropzoneComponent } from './Dropzone';
+
+export default { title: 'Components/Dropzone' };
+
+export const Dropzone = () => <DropzoneComponent />;
