+++ src/components/Upload/Reducer/index.ts
@@ -1,0 +1,4 @@
+export * from './uploadActions';
+export * from './uploadInitialState';
+export * from './uploadReducer';
+export * from './uploadTypings';
