+++ src/components/Upload/Reducer/uploadReducer.ts
