--- src/components/Upload/Upload.stories.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { Upload as UploadComponent } from './Upload';
-
-export default { title: 'Components/Upload' };
-
-export const Upload = () => <UploadComponent />;
