+++ src/components/Upload/UploadManager.tsx
@@ -1,0 +1,10 @@
+import React, { FC } from 'react';
+import { UploadState } from './Reducer';
+
+interface UploadManagerProps {
+  state: UploadState;
+}
+
+export const UploadManager: FC<UploadManagerProps> = () => {
+  return <div></div>;
+};
