+++ src/components/Upload/index.ts
@@ -1,2 +1,4 @@
-export * from './Upload';
-export * from './Context';
+export * from './Reducer';
+export * from './Dropzone';
+export * from './UploadContext';
+export * from './UploadManager';
